User Table Refactor
