+++ src/components/User/UserTable.js
@@ -1,16 +1,13 @@
 import {
-	Checkbox, Hidden, Paper, Table, TableBody, TableCell,
-	TableRow, withStyles, DialogTitle, Dialog, DialogContent,
-	DialogContentText, DialogActions, Button, Typography, IconButton,
+	Checkbox, Hidden, Table, TableBody, TableCell,
+	TableRow, withStyles, Typography,
 } from '@material-ui/core'
 import TC from 'components/Table/TC'
-import { Delete, Edit, PictureAsPdf, Add, StarBorder, Star } from 'variables/icons'
 import devicetableStyles from 'assets/jss/components/devices/devicetableStyles'
 import PropTypes from 'prop-types'
 import React, { Fragment } from 'react'
 import { withRouter } from 'react-router-dom'
 import EnhancedTableHead from 'components/Table/TableHeader'
-import EnhancedTableToolbar from 'components/Table/TableToolbar'
 import { ItemGrid, Info, Caption } from 'components'
 import { connect } from 'react-redux'
 import { pF, dateFormat } from 'variables/functions';
@@ -37,8 +34,7 @@ class UserTable extends React.Component {
 	
 	componentDidUpdate = () => {
 		if (this.props.saved === true) {
-			const { data } = this.props
-			const { selected } = this.state
+			const { data, selected } = this.props
 			let user = data[data.findIndex(d => d.id === selected[0])]
 			if (this.props.isFav({ id: user.id, type: 'user' })) {
 				this.props.s('snackbars.favorite.saved', { name: `${user.firstName} ${user.lastName}`, type: this.props.t('favorites.types.user') })
@@ -52,21 +48,7 @@ class UserTable extends React.Component {
 			}
 		}
 	}
-	handleToolbarMenuOpen = e => {
-		e.stopPropagation()
-		this.setState({ anchorElMenu: e.currentTarget })
-	}
-
-	handleToolbarMenuClose = e => {
-		e.stopPropagation();
-		this.setState({ anchorElMenu: null })
-	}
-
-	handleSearch = value => {
-		this.setState({
-			searchFilter: value
-		})
-	}
+	
 
 	handleRequestSort = (event, property) => {
 		this.props.handleRequestSort(event, property)
@@ -81,36 +63,6 @@ class UserTable extends React.Component {
 		}
 	}
 
-	handleSelectAllClick = (event, checked) => {
-		if (checked) {
-			this.setState({ selected: this.props.data.map(n => n.id) })
-			return;
-		}
-		this.setState({ selected: [] })
-	}
-
-	handleClick = (event, id) => {
-		event.stopPropagation()
-		const { selected } = this.state;
-		const selectedIndex = selected.indexOf(id)
-		let newSelected = [];
-
-		if (selectedIndex === -1) {
-			newSelected = newSelected.concat(selected, id);
-		} else if (selectedIndex === 0) {
-			newSelected = newSelected.concat(selected.slice(1))
-		} else if (selectedIndex === selected.length - 1) {
-			newSelected = newSelected.concat(selected.slice(0, -1))
-		} else if (selectedIndex > 0) {
-			newSelected = newSelected.concat(
-				selected.slice(0, selectedIndex),
-				selected.slice(selectedIndex + 1),
-			);
-		}
-
-		this.setState({ selected: newSelected })
-	}
-
 	handleChangePage = (event, page) => {
 		this.setState({ page });
 	}
@@ -119,100 +71,8 @@ class UserTable extends React.Component {
 		this.setState({ rowsPerPage: event.target.value })
 	}
 
-	handleDeleteUsers = async () => {
-		await this.props.handleDeleteUsers(this.state.selected)
-		this.setState({
-			selected: [],
-			anchorElMenu: null,
-			openDelete: false
-		})
-	}
-
-	handleOpenDeleteDialog = () => {
-		this.setState({ openDelete: true, anchorElMenu: null })
-	}
-
-	handleCloseDeleteDialog = () => {
-		this.setState({ openDelete: false })
-	}
-
-	isSelected = id => this.state.selected.indexOf(id) !== -1
-
-	handleEdit = () => {
-		this.props.history.push(`/management/user/${this.state.selected[0]}/edit`)
-	}
-	addToFav = (favObj) => {
-		this.props.addToFav(favObj)
-		this.setState({ anchorElMenu: null })
-	}
-	removeFromFav = (favObj) => {
-		this.props.removeFromFav(favObj)
-		this.setState({ anchorElMenu: null })
-	}
-	options = () => {
-		const { t, isFav, data } = this.props
-		const { selected } = this.state
-		let user = data[data.findIndex(d => d.id === selected[0])]
-		let favObj = {
-			id: user.id,
-			name: `${user.firstName} ${user.lastName}`,
-			type: 'user',
-			path: `/management/user/${user.id}`
-		}
-		let isFavorite = isFav(favObj)
-		return [
-			{ label: t('menus.edit'), func: this.handleEdit, single: true, icon: Edit },
-			{ label: isFavorite ? t('menus.favorites.remove') : t('menus.favorites.add'), icon: isFavorite ? Star : StarBorder, func: isFavorite ? () => this.removeFromFav(favObj) : () => this.addToFav(favObj) },
-			{ label: t('menus.exportPDF'), func: () => { }, icon: PictureAsPdf },
-			{ label: t('menus.delete'), func: this.handleOpenDeleteDialog, icon: Delete }
-		]
-	}
-
-	renderConfirmDelete = () => {
-		const { openDelete, selected } = this.state
-		const { data, t } = this.props
-
-		return <Dialog
-			open={openDelete}
-			onClose={this.handleCloseDeleteDialog}
-			aria-labelledby='alert-dialog-title'
-			aria-describedby='alert-dialog-description'
-		>
-			<DialogTitle id='alert-dialog-title'>{t('dialogs.delete.title.users')}</DialogTitle>
-			<DialogContent>
-				<DialogContentText id='alert-dialog-description'>
-					{t('dialogs.delete.message.users')}
-				</DialogContentText>
-				<div>
-					{selected.map(s => {
-						let u = data[data.findIndex(d => d.id === s)]
-						return <Info key={s}>&bull;{u.firstName + ' ' + u.lastName}</Info>})
-					}
-				</div>
-			</DialogContent>
-			<DialogActions>
-				<Button onClick={this.handleCloseDeleteDialog} color='primary'>
-					{t('actions.no')}
-				</Button>
-				<Button onClick={this.handleDeleteUsers} color='primary' autoFocus>
-					{t('actions.yes')}
-				</Button>
-			</DialogActions>
-		</Dialog>
-	}
-
-	addNewUser = () => { this.props.history.push('/management/users/new') }
-
-	renderTableToolBarContent = () => {
-		const { accessLevel } = this.props
-		let access = accessLevel.apiorg ? accessLevel.apiorg.edit ? true : false : false
-		return <Fragment>
-			{access ? <IconButton aria-label='Add new user' onClick={this.addNewUser}>
-				<Add />
-			</IconButton> : null
-			}
-		</Fragment>
-	}
+	isSelected = id => this.props.selected.indexOf(id) !== -1
+	
 	render() {
 		const { order, orderBy, data, classes, t } = this.props
 		const {  selected, rowsPerPage, page } = this.state
@@ -220,24 +80,14 @@ class UserTable extends React.Component {
 		if (data)
 			emptyRows = rowsPerPage - Math.min(rowsPerPage, data.length - page * rowsPerPage)
 		return (
-			<Paper className={classes.root}>
-
-				<EnhancedTableToolbar
-					anchorElMenu={this.state.anchorElMenu}
-					handleToolbarMenuClose={this.handleToolbarMenuClose}
-					handleToolbarMenuOpen={this.handleToolbarMenuOpen}
-					numSelected={selected.length}
-					options={this.options}
-					t={t}
-					content={this.renderTableToolBarContent()}
-				/>
+			<Fragment>
 				<div className={classes.tableWrapper}>
 					<Table className={classes.table} aria-labelledby='tableTitle'>
 						<EnhancedTableHead // ./ProjectTableHeader
 							numSelected={selected.length}
 							order={order}
 							orderBy={orderBy}
-							onSelectAllClick={this.handleSelectAllClick}
+							onSelectAllClick={this.props.handleSelectAllClick}
 							onRequestSort={this.handleRequestSort}
 							rowCount={data ? data.length : 0}
 							columnData={this.props.tableHead}
@@ -246,7 +96,7 @@ class UserTable extends React.Component {
 							customColumn={[ {
 								id: 'avatar', label: <div style={{ width: 40 }}/>
 							}, {
-								id: 'firstName', label: <Typography variant={'body1'}>Users</Typography>
+								id: 'firstName', label: <Typography paragraph classes={{ root: classes.paragraphCell + ' ' + classes.headerCell }}>Users</Typography>
 							}]}
 						/>
 						<TableBody>
@@ -265,7 +115,7 @@ class UserTable extends React.Component {
 										style={{ cursor: 'pointer' }}
 									>
 										<Hidden lgUp>
-											<TC checkbox content={<Checkbox checked={isSelected} onClick={e => this.handleClick(e, n.id)} />} />
+											<TC checkbox content={<Checkbox checked={isSelected} onClick={e => this.props.handleCheckboxClick(e, n.id)} />} />
 											<TC checkbox content={n.img ? <img src={n.img} alt='brken' className={classes.img} /> : <Gravatar default='mp' email={n.email} className={classes.img} />}/>
 
 											<TC content={
@@ -284,7 +134,7 @@ class UserTable extends React.Component {
 											}/>
 										</Hidden>
 										<Hidden mdDown>
-											<TC checkbox content={<Checkbox checked={isSelected} onClick={e => this.handleClick(e, n.id)}/>} />
+											<TC checkbox content={<Checkbox checked={isSelected} onClick={e => this.props.handleCheckboxClick(e, n.id)}/>} />
 											<TC checkbox content={n.img ? <img src={n.img} alt='brken' className={classes.img} /> : <Gravatar default='mp' email={n.email} className={classes.img} />} />
 											<TC FirstC label={`${n.firstName} ${n.lastName}`} />
 											<TC label={<a onClick={e => e.stopPropagation()} href={`tel:${n.phone}`}>{n.phone ? pF(n.phone, this.props.language) : n.phone}</a>} />
@@ -313,9 +163,7 @@ class UserTable extends React.Component {
 					handleChangePage={ this.handleChangePage }
 					handleChangeRowsPerPage={ this.handleChangeRowsPerPage }
 				/>
-				{this.renderConfirmDelete()}
-
-			</Paper>
+			</Fragment>
 		)
 	}
 }
