+++ src/views/Users/Users.js
@@ -1,32 +1,107 @@
 import React, { Component, Fragment } from 'react'
-import { withStyles } from '@material-ui/core';
+import { withStyles, Paper, IconButton, Dialog, DialogTitle, DialogContent, DialogContentText, DialogActions, Button } from '@material-ui/core';
 import projectStyles from 'assets/jss/views/projects';
 import UserTable from 'components/User/UserTable';
 import CircularLoader from 'components/Loader/CircularLoader';
 import GridContainer from 'components/Grid/GridContainer';
 import { deleteUser } from 'variables/dataUsers';
-import { People, Business } from 'variables/icons';
-import { filterItems, handleRequestSort } from 'variables/functions';
+import { People, Business, Add, Delete, PictureAsPdf, Edit, Star, StarBorder } from 'variables/icons';
+import { handleRequestSort } from 'variables/functions';
+import TableToolbar from 'components/Table/TableToolbar';
+import { Info } from 'components';
+import { customFilterItems } from 'variables/Filters';
 
 class Users extends Component {
 	constructor(props) {
 		super(props)
 
 		this.state = {
+			selected: [],
+			openDelete: false,
 			users: [],
 			userHeader: [],
+			anchorElMenu: null,
 			loading: true,
 			route: 0,
 			order: 'desc',
 			orderBy: 'firstName',
 			filters: {
 				keyword: '',
-				startDate: null,
-				endDate: null,
-				activeDateFilter: false
+				custom: []
 			}
 		}
 		props.setHeader('users.pageTitle', false, '', 'users')
+	}
+	ftUsers = () => {
+		const { t } = this.props
+		return [
+			{ key: 'firstName', name: t('users.fields.firstName'), type: 'string' },
+			{ key: 'lastName', name: t('users.fields.lastName'), type: 'string' },
+			{ key: 'email', name: t('users.fields.email'), type: 'string' },
+			{ key: 'org.name', name: t('orgs.fields.name'), type: 'string' }, 
+			// { key: 'devices[0].start', name: t('collections.fields.activeDeviceStartDate'), type: 'date' },
+			// { key: 'created', name: t('collections.fields.created'), type: 'date' },
+			// { key: 'activeDeviceStats.state', name: t('devices.fields.status'), type: 'dropDown', options: this.dLiveStatus() }
+			// { key: 'address', name: t('devices.fields.address'), type: 'string' },
+			// { key: 'locationType', name: t('devices.fields.locType'), type: 'dropDown', options: this.dLocationPlace() },
+			// { key: 'lat', name: t('calibration.stepheader.calibration'), type: 'diff', options: { dropdown: this.dCalibrated(), values: { false: [0] } } },
+		]
+	}
+	addFilter = (f) => {
+		let cFilters = this.state.filters.custom
+		let id = cFilters.length
+		cFilters.push({ ...f, id: id })
+		this.setState({
+			filters: {
+				...this.state.filters,
+				custom: cFilters
+			}
+		})
+		return id
+	}
+	removeFilter = (fId) => {
+		let cFilters = this.state.filters.custom
+		cFilters = cFilters.reduce((newFilters, f) => {
+			if (f.id !== fId) {
+				newFilters.push(f)
+			}
+			return newFilters
+		}, [])
+		this.setState({
+			filters: {
+				...this.state.filters,
+				custom: cFilters
+			}
+		})
+	}
+	handleEdit = () => {
+		this.props.history.push(`/management/user/${this.state.selected[0]}/edit`)
+	}
+	addToFav = (favObj) => {
+		this.props.addToFav(favObj)
+		this.setState({ anchorElMenu: null })
+	}
+	removeFromFav = (favObj) => {
+		this.props.removeFromFav(favObj)
+		this.setState({ anchorElMenu: null })
+	}
+	options = () => {
+		const { t, isFav, users } = this.props
+		const { selected } = this.state
+		let user = users[users.findIndex(d => d.id === selected[0])]
+		let favObj = {
+			id: user.id,
+			name: `${user.firstName} ${user.lastName}`,
+			type: 'user',
+			path: `/management/user/${user.id}`
+		}
+		let isFavorite = isFav(favObj)
+		return [
+			{ label: t('menus.edit'), func: this.handleEdit, single: true, icon: Edit },
+			{ label: isFavorite ? t('menus.favorites.remove') : t('menus.favorites.add'), icon: isFavorite ? Star : StarBorder, func: isFavorite ? () => this.removeFromFav(favObj) : () => this.addToFav(favObj) },
+			{ label: t('menus.exportPDF'), func: () => { }, icon: PictureAsPdf },
+			{ label: t('menus.delete'), func: this.handleOpenDeleteDialog, icon: Delete }
+		]
 	}
 	userHeader = () => {
 		const { t } = this.props
@@ -52,8 +127,10 @@ class Users extends Component {
 			// this.setState({ users: this.props.users })
 		}
 	}
+
 	filterItems = (data) => {
-		return filterItems(data, this.state.filters)
+		const { filters } = this.state
+		return customFilterItems(data, filters.custom)
 	}
 	handleRequestSort = (event, property, way) => {
 		let order = way ? way : this.state.order === 'desc' ? 'asc' : 'desc'
@@ -119,38 +196,157 @@ class Users extends Component {
 		this.setState({ loading: true })
 		await this.props.reload()
 	}
-	handleDeleteUsers = async (selected) => {
+	deleteUsers = async () => {
+		const { selected } = this.state
 		await selected.forEach(async u => {
 			await deleteUser(u)
 		})
-		await this.props.reload()
+		await this.reload()
 		this.snackBarMessages(1)
 	}
-	renderUsers = () => {
-		const { t } = this.props
-		const { loading, order, orderBy, filters, users } = this.state
-		return <GridContainer justify={'center'}>
-			{loading ? <CircularLoader /> : <UserTable
-				data={users}
-				tableHead={this.userHeader()}
-				handleFilterEndDate={this.handleFilterEndDate}
-				handleFilterKeyword={this.handleFilterKeyword}
-				handleFilterStartDate={this.handleFilterStartDate}
-				handleRequestSort={this.handleRequestSort}
-				handleDeleteUsers={this.handleDeleteUsers}
-				order={order}
-				orderBy={orderBy}
-				filters={filters}
-				t={t}
-			/>}
-		</GridContainer>
+	addNewUser = () => { this.props.history.push('/management/users/new') }
+
+	handleCheckboxClick = (event, id) => {
+		event.stopPropagation()
+		const { selected } = this.state;
+		const selectedIndex = selected.indexOf(id)
+		let newSelected = [];
+
+		if (selectedIndex === -1) {
+			newSelected = newSelected.concat(selected, id);
+		} else if (selectedIndex === 0) {
+			newSelected = newSelected.concat(selected.slice(1))
+		} else if (selectedIndex === selected.length - 1) {
+			newSelected = newSelected.concat(selected.slice(0, -1))
+		} else if (selectedIndex > 0) {
+			newSelected = newSelected.concat(
+				selected.slice(0, selectedIndex),
+				selected.slice(selectedIndex + 1),
+			);
+		}
+
+		this.setState({ selected: newSelected })
+	}
+	handleSelectAllClick = (event, checked) => {
+		if (checked) {
+			this.setState({ selected: this.props.users.map(n => n.id) })
+			return;
+		}
+		this.setState({ selected: [] })
+	}
+	handleDeleteUsers = async () => {
+		await this.deleteUsers()
+		this.setState({
+			selected: [],
+			anchorElMenu: null,
+			openDelete: false
+		})
 	}
 
+	handleOpenDeleteDialog = () => {
+		this.setState({ openDelete: true, anchorElMenu: null })
+	}
+
+	handleCloseDeleteDialog = () => {
+		this.setState({ openDelete: false })
+	}
+
+	renderConfirmDelete = () => {
+		const { openDelete, selected } = this.state
+		const { users, t } = this.props
+
+		return <Dialog
+			open={openDelete}
+			onClose={this.handleCloseDeleteDialog}
+			aria-labelledby='alert-dialog-title'
+			aria-describedby='alert-dialog-description'
+		>
+			<DialogTitle id='alert-dialog-title'>{t('dialogs.delete.title.users')}</DialogTitle>
+			<DialogContent>
+				<DialogContentText id='alert-dialog-description'>
+					{t('dialogs.delete.message.users')}
+				</DialogContentText>
+				<div>
+					{selected.map(s => {
+						let u = users[users.findIndex(d => d.id === s)]
+						return <Info key={s}>&bull;{u.firstName + ' ' + u.lastName}</Info>
+					})
+					}
+				</div>
+			</DialogContent>
+			<DialogActions>
+				<Button onClick={this.handleCloseDeleteDialog} color='primary'>
+					{t('actions.no')}
+				</Button>
+				<Button onClick={this.handleDeleteUsers} color='primary' autoFocus>
+					{t('actions.yes')}
+				</Button>
+			</DialogActions>
+		</Dialog>
+	}
+
+	renderUsers = () => {
+		const { t, classes } = this.props
+		const { loading, selected, order, orderBy, filters, users } = this.state
+		return <GridContainer justify={'center'}>
+			{loading ? <CircularLoader /> : <Paper className={classes.root}>
+				{this.renderConfirmDelete()}
+				<TableToolbar
+					ft={this.ftUsers()}
+					addFilter={this.addFilter}
+					removeFilter={this.removeFilter}
+					anchorElMenu={this.state.anchorElMenu}
+					handleToolbarMenuClose={this.handleToolbarMenuClose}
+					handleToolbarMenuOpen={this.handleToolbarMenuOpen}
+					numSelected={selected.length}
+					options={this.options}
+					t={t}
+					content={this.renderTableToolBarContent()}
+				/><UserTable
+					data={this.filterItems(users)}
+					selected={selected}
+					tableHead={this.userHeader()}
+					handleSelectAllClick={this.handleSelectAllClick}
+					handleFilterEndDate={this.handleFilterEndDate}
+					handleFilterKeyword={this.handleFilterKeyword}
+					handleFilterStartDate={this.handleFilterStartDate}
+					handleRequestSort={this.handleRequestSort}
+					handleDeleteUsers={this.handleDeleteUsers}
+					handleCheckboxClick={this.handleCheckboxClick}
+					order={order}
+					orderBy={orderBy}
+					filters={filters}
+					t={t}
+				/>
+			</Paper>}
+		</GridContainer>
+	}
+	handleToolbarMenuOpen = e => {
+		e.stopPropagation()
+		this.setState({ anchorElMenu: e.currentTarget })
+	}
+
+	handleToolbarMenuClose = e => {
+		e.stopPropagation();
+		this.setState({ anchorElMenu: null })
+	}
+	renderTableToolBarContent = () => {
+		const { accessLevel } = this.props
+		let access = accessLevel.apiorg ? accessLevel.apiorg.edit ? true : false : false
+		return <Fragment>
+			{access ? <IconButton aria-label='Add new user' onClick={this.addNewUser}>
+				<Add />
+			</IconButton> : null
+			}
+		</Fragment>
+	}
 	render() {
-		// const { users, filters } = this.state
+
 		return (
 			<Fragment>
+
 				{this.renderUsers()}
+				{/* </Paper> */}
 			</Fragment>
 		)
 	}
